Restructing to use engines. Adding examples.
Restructing to use an EngineLoader instead of the default Handlebars
template engine. Updated templates task to be called assemble and
use the EngineLoader for loading. A lot more work still needs to
be done, but it's a start.

Adding an example for using handlebars with yaml headers and a
layout.