Use `.template` to load deafult template types
Adding `defaultTemplates` to load `pages`, `partials`, and `layouts`.
Removing the methods specific for loading pages, partials and layouts and moving their documentation to `docs/api-default-templates.md`
I think this will be a good idea for any defaults that we're settings, then we can include the doc in the comments for the `default*` method.